// frontend/src/app/api/model/company.ts
/**
 * Atlas API
 *
 * Contact: TechSupport-ATLAS@sbb.ch
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Company { 
    /**
     * Company UIC Code
     */
    uicCode?: string;
    /**
     * Company Name
     */
    name?: string;
    /**
     * Company URL
     */
    url?: string;
    /**
     * Validity Start Date
     */
    startValidity?: Date;
    /**
     * Validity End Date
     */
    endValidity?: Date;
    /**
     * Company short name
     */
    shortName?: string;
    /**
     * Free text
     */
    freeText?: string;
    /**
     * Country Code (ISO)
     */
    countryCodeIso?: string;
}